+++ contrib/opencensus-ext-azure/opencensus/ext/azure/version.py
