+++ contrib/opencensus-ext-azure/setup.py
@@ -14,7 +14,7 @@
 
 from setuptools import find_packages
 from setuptools import setup
-from version import __version__
+from opencensus.ext.azure.version import __version__
 
 setup(
     name='opencensus-ext-azure',
